--- Assets/Scripts/Managers/UI/ResourceUIManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: db22ef3349c5f5540bc6e3ed1b4c13d5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
