+++ Assets/Scripts/Models/Gathering/ResourceManager.cs
@@ -1,9 +1,16 @@
 using System.Collections.Generic;
+using TMPro;
 using UnityEngine;
 
 public class ResourceManager : MonoBehaviour
 {
     public static ResourceManager Instance;
+    private static ResourceManager instance;
+
+    private static readonly Dictionary<Resource.Type, int> totalResources = new();
+
+    [SerializeField] private TextMeshProUGUI woodText;
+    public int startingWood;
 
     private void Awake()
     {
@@ -18,13 +25,28 @@ public class ResourceManager : MonoBehaviour
         }
     }
 
+    private void Start()
+    {
+        totalResources.Add(Resource.Type.Wood, startingWood);
+        totalResources.Add(Resource.Type.Stone, 0);
+        totalResources.Add(Resource.Type.Gold, 0);
+    }
+
+    public void Update()
+    {
+        woodText.text = totalResources[Resource.Type.Wood].ToString();
+    }
+
     public static void RegisterResource(Resource resource)
     {
-        ResourceUIManager.RegisterResource(resource);
+        totalResources[resource.type] += resource.amount;
     }
 
     public static bool RequestResource(Resource requestedResource)
     {
-        return ResourceUIManager.RequestResource(requestedResource);
+        if (totalResources[requestedResource.type] < requestedResource.amount) return false;
+
+        totalResources[requestedResource.type] -= requestedResource.amount;
+        return true;
     }
 }
